pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthclinclit"
version = "0.1.0"
description = "Download PMC full-text articles by DOI/PMID/PMCID and extract user-defined artefacts with an LLM."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Basel Al-Shaikhdeeb" }]
dependencies = [
    "httpx>=0.27",
    "lxml>=5.2",
    "pydantic>=2.7",
    "anthropic>=0.40",
    "typer>=0.12",
    "rich>=13.7",
    "tenacity>=8.3",
    "python-dotenv>=1.0",
]

[project.scripts]
synthclinclit = "synthclinclit.cli:app"

[tool.setuptools.packages.find]
include = ["synthclinclit*"]
